feat: 기물 움직임의 중복되는 코드 공통으로 묶어 추상화
