+++ src/main/java/chess/model/piece/King.java
@@ -1,11 +1,9 @@
 package chess.model.piece;
 
-import chess.model.Board;
 import chess.model.square.Direction;
-import chess.model.square.Square;
 import java.util.List;
 
-public class King extends Piece {
+public class King extends PointMovingPiece {
 
     public King(Color color) {
         super(color);
@@ -17,19 +15,7 @@ public class King extends Piece {
     }
 
     @Override
-    public boolean movable(Square source, Square target) {
-        return getDirection().stream()
-                .anyMatch(direction -> source.findLocation(direction, target));
-    }
-
-    @Override
-    public boolean isObstacleOnRoute(Board board, Square source, Square target) {
-        Piece targetPiece = board.get(target);
-        return isNotAlly(targetPiece);
-    }
-
-    @Override
-    List<Direction> getDirection() {
+    public List<Direction> getDirection() {
         return List.of(
                 Direction.EAST,
                 Direction.WEST,
