+++ src/main/java/chess/model/square/Direction.java
@@ -21,8 +21,6 @@ public enum Direction {
     WWN(-2, 1),
     WWS(-2, -1);
 
-    private static final String INVALID_DIRECTION_ERROR_MESSAGE = "유효하지 않은 방향입니다.";
-
     private final int row;
     private final int col;
