+++ src/test/java/chess/model/ColorTest.java
@@ -11,7 +11,7 @@ public class ColorTest {
 
     @ParameterizedTest
     @DisplayName("진영의 색깔이 검은색인지 확인")
-    @CsvSource(value = {"WHITE:false", "NOTHING:false", "BLACK:true"}, delimiter = ':')
+    @CsvSource(value = {"WHITE:false", "EMPTY:false", "BLACK:true"}, delimiter = ':')
     void isBlack(Color color, boolean isBlack) {
         assertThat(color.isBlack()).isEqualTo(isBlack);
     }
